test : 쉐어 추천 API 실패케이스 통합 테스트 구현
